lab6/choreography/user/requirements.txt
google-cloud-pubsub==2.16.0
schedule